nousr: minimal implementation
The following libc calls are intercepted :

   __xstat, __lxstat, open, open64, access

The code currently considers the following variables :

  TMPDIR : is handled like /tmp (always allowed)
  NOUSR_PRJBASE : always allowed
  NOUSR_TCBASE  : always allowed, limits toolchain match to this path

If both NOUSR_PRJBASE and NOUSR_TCBASE are set, then the filtering is
restricted by default and does not allow access outside the permitted
paths. This will not always work and is basically usable only for
debugging.