nousr: fix the is_path_ok() logic :
  - allow /dev just like /tmp (eg: /dev/null, /dev/random, ...)
  - perform these tests only if needed (ie: if both tcbase and prjbase
    are set)
  - always allow relative paths even when prjbase and tcbase are set

Some corner cases remain in strict mode. For example, for an unknown
reason, when "mips-24kc_gcc47l_glibc218-linux-gnu-gcc" is called, it
properly finds that it's under /f/tc/mips/... but still it performs
some access() calls to check for the same binary in the PATH, then it
does another one using xstat().