nousr: add an extra NOUSR_STRICT variable
This variable enables or disable the strict check based on NOUSR_TCBASE
and NOUSR_PRJBASE. When all three variables are set, then no file access
to absolute paths outside of NOUSR_TCBASE, NOUSR_PRJBASE, /dev and /tmp
is permitted. It can be used for debugging, but will sometimes break
even on valid setups especially if some symlinks are used in the build
environment.