nousr: add a trace option with NOUSR_TRACE
NOUSR_TRACE may be used to force called programs to have their names and
arguments displayed when started.  This can be useful to trace makefiles.
This variable takes an integer value.  The first one, 1 only enables tracing
of intercepted programs. Value 2 goes a bit further and reports any program.
Each line is reported on stderr, prefixed with "##nousr##" so that they are
easily greppable.