nousr: panic in case a syscall is not defined.
In order to help with debugging, this emits an error message and exists
if a program uses a syscall which is null. This should never happen but
was caught a few times already on some distros, possibly indicating that
some memory was overwritten. It's better to report the problem loudly
than crashing in a crypting segfault.