nousr: do not try to dereference _dl_argv[] if it's null
Apparently _dl_argv[] is not always present when in the constructor.
It might be possible that it sometimes gets updated later or that
there is a good reason for it to be null sometimes. Let's protect
against this by simply not emitting the command line arguments when
not set. That's the only use we have for it so it's not critical.