nousr: reopen stderr if not yet opened
Depending on the program's load order we may be called before stderr is
assigned, which segfaults. Let's do it ourselves in such a case, using
fdopen(). In case it still fails, we emit an error on fd=2 and return
or exit depending on what's best where we can do it.